s2wasm: add outfile
Use some C++ in a few places. I'll propagate similar changes to the rest of the codebase later. I also need to turn off colors when outputting to a file (isatty on stdout doesn't do that with -o) but I'll do it in a separate PR because it'll touch more files.