Add an option to  allocate space for user stack at link time
Currently s2wasm pre-allocates space for the special __stack_pointer
variable but not for the actual stack, and __stack_pointer is null. On
real systems the stack is allocated by the dynamic linker, thread
runtime, or other toolchain/system-provided startup code. Since we don't
have any of that in wasm yet, just add an option to allocate the stack
in the linker for now, so we can continue to run single-file libc-free tests.