Support start
As spec'd in: https://github.com/WebAssembly/design/pull/495
And discussed in: https://github.com/WebAssembly/spec/issues/231

This will make it simpler and more uniform to add a start entry point.

s2wasm is the right place to add start because it'll eventually need to
do other basic setup, e.g. put code in start to setup the stack, as
dschuff is doing in:
  https://github.com/WebAssembly/binaryen/pull/179
Or rather, the linker is the right place and s2wasm happens to act as
our linker right now.