Generate emscripten dynCall thunks in s2wasm
Part of the usual emscripten glue for asm.js is thunks for calling
functions on the module given a function pointer; aka a dynCall. This is
implemented via thunks generated on the Module called
'dynCall_<signature>'. This gives support to s2wasm to generate these
thunks in the wasm module along with other emscripten glue. It also
introduces a flag to s2wasm to make generation of emscripten glue
optional.