ubsan: only recover from some of the failures
I fixed a few of the existing ubsan failures as part of #404, but would
like to prevent regressions. This patch allows some of the sanitizer
checks to continue execution (null, shift, signed-integer-overflow) and
makes all the other ones hard errors (preventing regressions).

Once the remaining errors are addressed we should commit #404 which
moves ubsan to -fno-sanitize-recover=all. This likely requires a good
checked-math library.