Generate thunks for address-taken imports
Under emscripten, C code can take the address of a function implemented
in Javascript (which is exposed via an import in wasm). Because imports
do not have linear memory address in wasm, we need to generate a thunk
to be the target of the indirect call; it call the import directly.

This is facilited by a new .s directive (.functype) which declares the
types of functions which are declared but not defined.

Fixes https://github.com/WebAssembly/binaryen/issues/392