Fix Appveyor badge URL to be unambiguous (#626)
The URL using the name of the GitHub project only works if no other appveyor project is using that github repo; that is apparently not true now. So use the unique URL for the WebAssembly appveyor account's project, and use the status of the master branch instead of the global status.