Fix a parsing bug introduced by #615
Name lhs = getStrToSep();
if (!skipEqual()){
  s = strchr(s, '\n');
  if (!s) break;
  continue;
}

The above code snippet introduced by #615 has a bug when there is only
one word (e.g. ".text") in a line. If there is only one word in a line,
skipEqual() also skips the newline character at the end of the line, and
strchr(s, '\n') moves the cursor to the end of the next line,
effectively skipping the whole next line.