Add a test case
In this test case, there is a '.text' directive, which is only one word,
before '.type foo,@function' line.
```
.text
.type  foo,@function
```

Because of the bug mentioned in the commit, s2wasm used to produce code
like this:
```
(import $foo "env" "foo")
(func $main (result i32)
  (call_import $foo)
  (i32.const 0)
)
(func $foo
)
```
which is wrong, because it tries to import foo and call it with
call_import even though there is already foo function implemented in the
module. After this patch, the right result should be:
```
(func $main (result i32)
  (call $foo)
  (i32.const 0)
)
(func $foo
)
```