Create the dummy nullptr function if there are any call-indirects
Previously it was only created if there were any address-taken
functions. However new module validation rules require the existence of
a table for any call-indirects to validate (even if they are dead and
never called). However this use case seems common enough that we might
want to make it continue to work. So create the dummy-nullptr function
if there are any call_indirect instructions even if there are no
address-taken functions. (They will trap at runtime if called, as before).