Update waterfall to 11453 and re-enable torture tests
Torture tests from the clang were disabled with the first 0xc
udpate. This PR updates the tests to a new waterfall build and turns
them on (also updating the known failures for 0xc). It also updates the
download URL since mac and windows bots have been added to the
waterfall (they are not up yet and support in this script is not tested
yet).