Let s2wasm emit binary output (#1465)
* Emit invokeFuncs list as metadata

* Refactor s2wasm to use ModuleWriter

* Fix wasm-emscripten-finalize metadata output for binary output

* Add a flag to emit binary from s2wasm

NOTE: I chose to emit text by default, and binary behind a flag. This
mismatches with asm2wasm (and the expectations of users of a "2wasm"
tool), but doesn't break any existing users of s2wasm. If s2wasm is
deprecated in favor of lld, this will be the least disruptive change,
and we won't have to live with awkward defaults for too long.

* Emit source maps in the binary output of s2wasm

* Only emit binary with an outfile specified