include the filename in erroneous diff output where possible (#1452)
Failing test cases often start out with:

```
incorrect output, diff:

--- expected
+++ actual
```

which makes it difficult to figure out where the expected output might
live.  That information can be derived from examining the tests, of
course, but it'd be much nicer if it were provided in the diff to see
straightaway.

We do this by introducing a new check, one which takes a filename of
expected output, which then enables us to display the failing file,
e.g.:

```
incorrect output, diff:

--- /home/froydnj/src/binaryen.git/test/passes/code-folding.txt
+++ actual
```

which is arguably nicer.  Having this new check also enables reducing
some boilerplate `open(...).read()` calls in various places.

There are still a few places using `fail_if_not_identical`, usually
because `.strip()` is used on the expected output.