reorder locals in wasm2asm (#1482)
The documentation for the simplify-locals pass suggests running
reorder-locals after it to clean up unnecessary locals.  wasm2asm wasn't
doing this, which meant that generated code had a number of unused
variables.  A good minimizer will probably clean that up, but let's go
ahead and clean it up in wasm2asm itself.