remap {get,set}_local indices (#1486)
When lowering i64 values in a function, we create new local variables
for all of the i64 local variables, one local for the low bits, and one
for the high bits.  We create a mapping between the old locals and the
new as well.  During translation, when we encountered a `get_local` that
didn't have type `i64`, we skipped it, on the supposition that there was
nothing to do.  But that's not true; the local it was getting may have
been remapped to a new index in the lowered function, and we need to
account for that change.  Similar logic holds for `set_local`.