improve remove-unused-module-elements (#1532)
Remove the entire memory/table when possible, in particular, when not imported, exported, or used. Previously we did not look at whether they were imported, so we assumed we could never remove them.

Also add a variant that removes everything but functions, which can be useful when reducing a testcase that only cares about code in functions.