wasm2asm: Finish f32/f64 operations
This commit finishes the f32 and f64 operations in the wasm2asm translation
tool. The spec test for `f32.wast` and `f64.wast` are now enabled and should be
checked on PRs from now on.

Along the way this also enables a number of other spec tests that should now be
passing, and a summary of all the changes are:

* The `f32.nearest` operation (and f64) was implemented in a lowering pass.
  Previously this erroneously called the `Math.NEAREST` function (which doesn't
  exist, nor does `Math.nearest`).
* The `f32.trunc` operation (and f64) was also implemented in a lowering pass.
  Previously the lowering called `Math.trunc` which wasn't valid asm.js
* Some functions injected in `wasm2asm.h` were moved to the new
  `RemoveNonAsmOps` pass so they can run a little more aggressively through
  optimizations as well as hopefully get all contained in one location.
* Optimizations were turned on by default for `wasm2asm` so the code generated
  by the `RemoveNonAsmOps` pass (as well as i64 -> i32) can get cleaned up.
* A `flatten` pass was inserted just before i64 -> i32 to hack around #1480